Banner Image and welcome text
In this commit I added a banner image to the page, along with some welcome text to overlay the banner. I added a short blurb about Caffeine Dreams and a few pictures to round it out. It is simple, but I do not want to go overboard at this point. For now it is enough to allow me to move on to the other pages.